Move not main files to /lib/
