+++ lib/command.js
