+++ lib/commandlist.js
