+++ lib/message.js
