+++ lib/protocol.js
