+++ lib/room.js
