+++ lib/roomlist.js
