+++ lib/song.js
