+++ lib/typecheck.js
